+++ examples/node/basic_js_node_7_up/example.js
@@ -10,20 +10,21 @@ const apiKey = process.env.API_KEY;
 
 const run = async () => {
   const params = {
+    engine: "google",
     q: "Coffee",
     api_key: apiKey,
   };
 
   // Show result as JSON (async/await)
-  const response1 = await getJson("google", params);
+  const response1 = await getJson(params);
   console.log(response1["organic_results"]);
 
   // Show result as JSON (callback)
-  getJson("google", params, (json) => console.log(json["organic_results"]));
+  getJson(params, (json) => console.log(json["organic_results"]));
 
   // Use global config
   config.api_key = apiKey;
-  const response2 = await getJson("google", { q: "Coffee" });
+  const response2 = await getJson({ engine: "google", q: "Coffee" });
   console.log(response2["organic_results"]);
 };