+++ examples/node/pagination_js_node_7_up/example.js
@@ -12,12 +12,13 @@ const extractLinks = (results) => results.map((r) => r.link);
 
 const run = async () => {
   const params = {
+    engine: "google",
     q: "Coffee",
     api_key: apiKey,
   };
 
   // Pagination (async/await)
-  let page1 = await getJson("google", params);
+  let page1 = await getJson(params);
   console.log(
     "First page links",
     extractLinks(page1.organic_results)
@@ -31,7 +32,7 @@ const run = async () => {
   }
 
   // Pagination (callback)
-  getJson("google", params, (page1) => {
+  getJson(params, (page1) => {
     console.log(
       "First page links",
       extractLinks(page1.organic_results)
@@ -48,18 +49,16 @@ const run = async () => {
 
   // Use global config
   config.api_key = apiKey;
-  page1 = await getJson("google", { q: "Coffee" });
-  if (page1.next) {
-    page2 = await page1.next();
-    console.log(
-      "Second page links",
-      extractLinks(page2.organic_results)
-    );
-  }
+  page1 = await getJson({ engine: "google", q: "Coffee" });
+  page2 = await page1.next?.();
+  console.log(
+    "Second page links",
+    extractLinks(page2?.organic_results),
+  );
 
   // Pagination loop (async/await)
   let links = [];
-  let page = await getJson("google", { q: "Coffee" });
+  let page = await getJson({ engine: "google", q: "Coffee" });
   while (page) {
     links.push(...extractLinks(page.organic_results));
     if (links.length >= 30) break;
@@ -69,7 +68,7 @@ const run = async () => {
 
   // Pagination loop (callback)
   links = [];
-  getJson("google", { q: "Coffee" }, (page) => {
+  getJson({ engine: "google", q: "Coffee" }, (page) => {
     links.push(...extractLinks(page.organic_results));
     if (links.length < 30 && page.next) {
       page.next();