+++ src/utils.ts
@@ -1,9 +1,9 @@
-import type { EngineName, EngineParameters } from "./types.ts";
+import type { EngineParameters } from "./types.ts";
 import { version } from "../version.ts";
 import https from "node:https";
 import qs from "node:querystring";
-import url from "node:url";
 import { RequestTimeoutError } from "./errors.ts";
+import { EngineMap } from "./engines/engine_map.ts";
 
 /**
  * This `_internals` object is needed to support stubbing/spying of
@@ -22,7 +22,7 @@ function getBaseUrl() {
   return "https://serpapi.com";
 }
 
-type NextParameters<E extends EngineName = EngineName> = {
+type NextParameters<E extends keyof EngineMap> = {
   [
     K in keyof Omit<
       EngineParameters<E>,
@@ -30,22 +30,16 @@ type NextParameters<E extends EngineName = EngineName> = {
     >
   ]: string;
 };
-export function extractNextParameters<E extends EngineName = EngineName>(
-  json: {
-    serpapi_pagination?: { next: string };
-    pagination?: { next: string };
-  },
-) {
+export function extractNextParameters<E extends keyof EngineMap>(json: {
+  serpapi_pagination?: { next: string };
+  pagination?: { next: string };
+}) {
   const nextUrlString = json["serpapi_pagination"]?.["next"] ||
     json["pagination"]?.["next"];
 
   if (nextUrlString) {
-    const nextUrl = new url.URL(nextUrlString);
-    const nextParameters: Record<string, string> = {};
-    for (const [k, v] of nextUrl.searchParams.entries()) {
-      if (k === "engine") continue;
-      nextParameters[k] = v;
-    }
+    const nextUrl = new URL(nextUrlString);
+    const nextParameters = Object.fromEntries(nextUrl.searchParams.entries());
     return nextParameters as NextParameters<E>;
   }
 }