+++ CHANGELOG.md
@@ -10,8 +10,8 @@ and this project adheres to
 
 ### Added
 
-- Expose `EngineName`, `EngineParameters` and `AllowArbitraryParams` types.
-- Add support for Node.js 7.10.1 and newer.
+- Expose `EngineParameters` type.
+- Expose `InvalidArgumentError` error.
 
 ### Changed